[2.0] Protocol lib

+++ src/main/java/cn/nukkit/block/BlockBricksEndStone.java
@@ -19,12 +19,12 @@ public class BlockBricksEndStone extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
-        return 0.8;
+    public float getHardness() {
+        return 0.8f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 4;
     }
