+++ src/main/java/cn/nukkit/block/BlockBricksNether.java
@@ -23,12 +23,12 @@ public class BlockBricksNether extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
+    public float getHardness() {
         return 2;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 10;
     }
