+++ src/main/java/cn/nukkit/block/BlockDriedKelp.java
@@ -9,13 +9,13 @@ public class BlockDriedKelp extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
+    public float getHardness() {
         return 0.5f;
     }
 
     @Override
-    public double getResistance() {
-        return 2.5;
+    public float getResistance() {
+        return 2.5f;
     }
 
     @Override
