+++ src/main/java/cn/nukkit/block/BlockHugeMushroomBrown.java
@@ -23,12 +23,12 @@ public class BlockHugeMushroomBrown extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
-        return 0.2;
+    public float getHardness() {
+        return 0.2f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 1;
     }
