+++ src/main/java/cn/nukkit/block/BlockMagma.java
@@ -22,12 +22,12 @@ public class BlockMagma extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
-        return 0.5;
+    public float getHardness() {
+        return 0.5f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 30;
     }
