+++ src/main/java/cn/nukkit/block/BlockSand.java
@@ -18,13 +18,13 @@ public class BlockSand extends BlockFallable {
     }
 
     @Override
-    public double getHardness() {
-        return 0.5;
+    public float getHardness() {
+        return 0.5f;
     }
 
     @Override
-    public double getResistance() {
-        return 2.5;
+    public float getResistance() {
+        return 2.5f;
     }
 
     @Override
