+++ src/main/java/cn/nukkit/block/BlockSnow.java
@@ -13,12 +13,12 @@ public class BlockSnow extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
-        return 0.2;
+    public float getHardness() {
+        return 0.2f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 1;
     }
