+++ src/main/java/cn/nukkit/block/BlockStairsAndesite.java
@@ -9,17 +9,17 @@ import cn.nukkit.utils.Identifier;
  */
 public class BlockStairsAndesite extends BlockStairs {
 
-    public BlockStairsAndesite(Identifier id){
+    public BlockStairsAndesite(Identifier id) {
         super(id);
     }
 
     @Override
-    public double getHardness(){
-        return 1.5;
+    public float getHardness() {
+        return 1.5f;
     }
 
     @Override
-    public double getResistance(){
+    public float getResistance() {
         return 30;
     }
