+++ src/main/java/cn/nukkit/block/BlockStairsDiorite.java
@@ -9,17 +9,17 @@ import cn.nukkit.utils.Identifier;
  */
 public class BlockStairsDiorite extends BlockStairs {
 
-    public BlockStairsDiorite(Identifier id){
+    public BlockStairsDiorite(Identifier id) {
         super(id);
     }
 
     @Override
-    public double getHardness(){
-        return 1.5;
+    public float getHardness() {
+        return 1.5f;
     }
 
     @Override
-    public double getResistance(){
+    public float getResistance() {
         return 30;
     }
