+++ src/main/java/cn/nukkit/block/BlockStairsPrismarine.java
@@ -9,17 +9,17 @@ import cn.nukkit.utils.Identifier;
  */
 public class BlockStairsPrismarine extends BlockStairs {
 
-    public BlockStairsPrismarine(Identifier id){
+    public BlockStairsPrismarine(Identifier id) {
         super(id);
     }
 
     @Override
-    public double getHardness(){
-        return 1.5;
+    public float getHardness() {
+        return 1.5f;
     }
 
     @Override
-    public double getResistance(){
+    public float getResistance() {
         return 30;
     }
