+++ src/main/java/cn/nukkit/level/generator/Generator.java
@@ -4,10 +4,9 @@ import cn.nukkit.level.ChunkManager;
 import cn.nukkit.level.chunk.Chunk;
 import cn.nukkit.level.chunk.IChunk;
 import cn.nukkit.math.BedrockRandom;
-import cn.nukkit.math.Vector3f;
+import com.nukkitx.math.vector.Vector3i;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.util.Map;
 
 @NotThreadSafe
 public interface Generator {
@@ -22,5 +21,5 @@ public interface Generator {
 
     String getSettings();
 
-    Vector3f getSpawn();
+    Vector3i getSpawn();
 }
